fix: build error
Refs: #18
