+++ src/backend/api/internal/models/common.go
@@ -1,0 +1,11 @@
+package models
+
+// PaginationInfo represents common pagination metadata
+type PaginationInfo struct {
+	CurrentPage int  `json:"current_page"`
+	PageSize    int  `json:"page_size"`
+	TotalItems  int  `json:"total_items"`
+	TotalPages  int  `json:"total_pages"`
+	HasNext     bool `json:"has_next"`
+	HasPrev     bool `json:"has_prev"`
+}
