+++ src/backend/api/internal/models/property.go
@@ -129,11 +129,3 @@ type BuildingListResponse struct {
 }
 
 // PaginationInfo contains pagination metadata
-type PaginationInfo struct {
-	CurrentPage int  `json:"current_page"`
-	PageSize    int  `json:"page_size"`
-	TotalItems  int  `json:"total_items"`
-	TotalPages  int  `json:"total_pages"`
-	HasNext     bool `json:"has_next"`
-	HasPrev     bool `json:"has_prev"`
-}
